requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
PyJWT>=2.7
drf-yasg>=1.21.6
django-cors-headers>=4.0.0
